+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 - Fixes an issue where the main content section wasn't correctly flexed - [#91](https://github.com/PrefectHQ/miter-design/pull/91)
 - Make sure input component emits native keypresses - [#90](https://github.com/PrefectHQ/miter-design/pull/90)
+- Add Time Picker component - [#92](https://github.com/PrefectHQ/miter-design/pull/92)
+- Add Date Picker component - [#39](https://github.com/PrefectHQ/miter-design/pull/39)
 - Fix input styling on FF and position: absolute for Radio Buttons - [#89](https://github.com/PrefectHQ/miter-design/pull/89)
 - Fix bug where selecor menus were below other relative elements - [#88](https://github.com/PrefectHQ/miter-design/pull/88)
 - Allow customizable height/width on selectors - [#88](https://github.com/PrefectHQ/miter-design/pull/88)